- added essentiapool->weka ARFF converter - added weka attribute selector results to notebook - todo: add use weka to find octave error attributes

+++ pitchEst/notebook/util.py
@@ -9,10 +9,10 @@ def normalise(x):
         return x
 
 def freq2midi(freq, fref=440.):
-    return np.log2(freq / float(fref)) * 12 + 49
+    return (np.log2(np.float64(freq) / np.float64(fref)) * 12.) + 49
 
 def midi2freq(midi, fref=440.):
-    return 2**((midi - 49) / 12.) * fref
+    return 2**((np.float64(midi) - 49) / 12.) * fref
